Linux support! Also a bunch of fixes and performance improvements
This little project just came to mind recently, and i remember it
being heavily dependent on Windows APIs, despite being possible
to port it to linux. So here it is, albeit without audio for now.

Since things just run way too fast now due to the massive
optimizations done to the game logic, a frame limiter had to be
included as well.

Still, there might be new bugs introduced by this commit, as it is
VERY large and touches up on pretty much everything... something to
be looked at later.