Gigantic commit with audio and gameplay improvements
Also untested on windows, but so far, it's pretty decent on linux,
marking a massive improvement on the audio subsystem. Now SFX and
BGM can be played at the same time and without the need of using
the device's buzzer, courtesy of miniaudio. This also comes with a
ton of new sounds to boot.

BGMs are copyright free, but still have their authors credited on
the media info (even those that didn't have any info previously).

SFX were made by me.

Besides that, combat was heavily reworked and is FAR more engaging
now. Same for menus... although there's lots of unfinished stuff and
bugs to sort out, for example more backgrounds, reimplementing the
first boss and everything after (maybe expanding beyond the initial
scope?).

I think this can become something worthwhile, so i'll keep working
on it.